[AN] refactor(#488): todo/progress 관련 DTO 수정
