--- android/gitudy/data/src/main/java/com/takseha/data/dto/mystudy/MyStudyWithTodo.kt
@@ -1,12 +1,0 @@
-package com.takseha.data.dto.mystudy
-
-import com.takseha.data.dto.feed.StudyInfo
-
-data class MyStudyWithTodo(
-    val studyInfo: StudyInfo,
-    val todoTitle: String?,
-    val todoTime: String?,
-    val todoCheck: TodoStatus?,
-    val todoCheckNum: Int?
-)
-
