+++ android/gitudy/data/src/main/java/com/takseha/data/dto/mystudy/TodoListResponse.kt
@@ -9,24 +9,3 @@ data class TodoListResponse(
     @SerializedName("cursor_idx")
     val cursorIdx: Int
 )
-
-data class Todo(
-    @SerializedName("detail")
-    val detail: String = "",
-    @SerializedName("id")
-    val id: Int = 0,
-    @SerializedName("study_info_id")
-    val studyInfoId: Int = 0,
-    @SerializedName("title")
-    val title: String = "",
-    @SerializedName("todo_date")
-    val todoDate: String = "",
-    @SerializedName("todo_code")
-    val todoCode: String = "",
-    @SerializedName("todo_link")
-    val todoLink: String = "",
-    @SerializedName("created_date_time")
-    val todoSetDate: String = "",
-    @SerializedName("commits")
-    val commitList: List<Commit> = listOf()
-)
