+++ android/gitudy/data/src/main/java/com/takseha/data/dto/mystudy/TodoProgressResponse.kt
@@ -6,8 +6,8 @@ import com.google.gson.annotations.SerializedName
 data class TodoProgressResponse(
     @SerializedName("complete_member_count")
     val completeMemberCount: Int,
-    @SerializedName("todo_id")
-    val todoId: Int,
+    @SerializedName("todo")
+    val todo: Todo,
     @SerializedName("total_member_count")
     val totalMemberCount: Int
 )
