+++ android/gitudy/presentation/src/main/res/layout/activity_my_study_main.xml
@@ -315,27 +315,6 @@
                                 android:layout_marginRight="4dp"
                                 android:textColor="@color/BASIC_BLUE" />
                         </LinearLayout>
-
-                        <LinearLayout
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content">
-
-                            <TextView
-                                style="@style/B7_2"
-                                android:layout_width="wrap_content"
-                                android:layout_height="wrap_content"
-                                android:layout_marginRight="4dp"
-                                android:text="ToDoCode"
-                                android:textColor="@color/GS_500" />
-
-                            <TextView
-                                android:id="@+id/todoCode"
-                                style="@style/B7_2"
-                                android:layout_width="wrap_content"
-                                android:layout_height="wrap_content"
-                                android:layout_marginRight="4dp"
-                                android:textColor="@color/BLACK" />
-                        </LinearLayout>
                     </LinearLayout>
 
 
@@ -343,7 +322,7 @@
                         android:id="@+id/todoDetailBody"
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content"
-                        android:layout_marginTop="25dp"
+                        android:layout_marginTop="15dp"
                         android:visibility="visible"
                         app:layout_constraintEnd_toEndOf="parent"
                         app:layout_constraintStart_toStartOf="parent"
@@ -354,7 +333,6 @@
                             style="@style/B7_2"
                             android:layout_width="match_parent"
                             android:layout_height="wrap_content"
-                            android:text="@string/study_to_do_body"
                             android:textColor="@color/GS_500" />
                     </LinearLayout>
                 </androidx.constraintlayout.widget.ConstraintLayout>
